telegram-plex-bot/requirements.txt
python-telegram-bot
requests
qbittorrent-api
plexapi
python-dotenv
rarfile
patool